refactor: make sub packages in domain package
