+++ src/main/java/nextstep/subway/domain/repository/LineRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.repository;
 
+import nextstep.subway.domain.entity.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
