+++ src/main/java/nextstep/subway/domain/repository/StationRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.repository;
 
+import nextstep.subway.domain.entity.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
